Fixed a bug in SpriteObject -> GetFrameIndexForDeltaTime()
When a frame takes longer than the duration of a full infinite loop sequence, sprites would get stuck.
I modified the for loop to keep iterating until deltaTime is bigger than fTimeArrayCachedNextFrameTime again.
This will also give it a more consistent result, since it can now return an index higher then effectiveNumFrames (same as when not using a 'time' array)