Added View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION flag to immersive & immersiveSticky modes.
This will prevent the resize event on Android which happens when starting the application. The application will immediately have the right dimension.
Without this, a Corona app on a 1920x1080 screen will start at 1776x1080 and resizes to 1920x1080 shortly after starting up (a "resize" event occurs)
With this, a Corona app will immediately start with 1920x1080 as display.pixelWidth & display.pixelHeight