Allow developers to use complex settings files
Attempting to load a more complex config.lua will raise a warning and attempting to do the same with build.settings seems to result in the file not being loaded entirely.

This "complex file" may be attributed to loading a single (configuration) file in either file, so allow the developers to explicitly set their settings as a `complexFile` to allow them to be validated as normal.